add users to post model and templates

+++ src/posts/views.py
@@ -1,6 +1,6 @@
 from django.contrib import messages
 from django.shortcuts import render, get_object_or_404, redirect
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 
 # Create your views here.
@@ -9,11 +9,14 @@ from .forms import PostForm
 
 
 def post_create(request):
+    if not request.user.is_authenticated():
+        raise Http404
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
         # print(form.cleaned_data)
         # print(form.cleaned_data.get('title'))
+        instance.user = request.user
         instance.save()
         messages.success(request, "Successfully Created")
         return HttpResponseRedirect(instance.get_absolute_url())
@@ -49,10 +52,13 @@ def post_list(request):
 
 
 def post_update(request, slug=None):
+    if not request.user.is_authenticated():
+        raise Http404
     instance = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
         instance = form.save(commit=False)
+        instance.user = request.user
         instance.save()
         messages.success(request, "Successfully Saved")
         return HttpResponseRedirect(instance.get_absolute_url())
@@ -63,6 +69,8 @@ def post_update(request, slug=None):
 
 
 def post_delete(request, slug):
+    if not request.user.is_authenticated():
+        raise Http404
     instance = get_object_or_404(Post, slug=slug)
     instance.delete()
     messages.success(request, "Successfully Deleted")
